created rest api for user database and clean up empty/unneccessary lines
